fixes enemy warping
NOTES
related to issue #17

https://github.com/misael-diaz/sonic-infinite-scroller-game-dev/issues/17

zeros the x-component of the velocity of enemies that have exploded
it makes sense that they cannot move any longer at least in the
x-direction

improves the clamping of enemy coordinates by traversing the
game floor (all the platforms that comprise the floor); we ignore
the warp platforms `eta` and `rho` because we don't want enemies
to stay stuck there

we have tested the clamping with enemies movies with a faster
velocity and it works as expected

the previous algorithm failed because it assumed continuity of
the game floor and that's only true at the beginning before any
warping happens

the traversal of platforms that conform the floor happens is done
in their x-y sort ordering and takes into account if the enemy
is moving to the left or right; once we have determined the platform
that is next to the "void" we set the min/max coordinate along the
x-axis for the enemy. Note that we could have used MIN/MAX functions
instead of clamping because we only need to fix the x coordinate
according to the direction of movement of the enemy. For instance
if the enemy is moving to the right we care to fix the max position
the min position could be zero and it wouldn't affect anything.

to implement this note that we stored the position of the platform in
the sorted-platform list to save us the trouble from doing a search
for it; the property `platfno` was not used for platforms so we did
not have to invest more space to do this so it was free in terms of
space