adds wall jumping
NOTES

related to tasks #33, #35 , #41 and #42

https://github.com/misael-diaz/sonic-infinite-scroller-game-dev/issues/33
https://github.com/misael-diaz/sonic-infinite-scroller-game-dev/issues/35
https://github.com/misael-diaz/sonic-infinite-scroller-game-dev/issues/41
https://github.com/misael-diaz/sonic-infinite-scroller-game-dev/issues/42

TODO
shift the block at the eta platform upward so that sonic won't get
trapped; this is a minor issue and so it's worth to commit the
work done thus far

- adds code to release the block on sonic (by means of springing)
- renames block `beta` to `eta` because it's linked to the eta platform
- renames `en_check_overlapped` to `en_check_overlapping`
- `en_check_overlapping` does its job when the entity is at the floor level
- extends `en_check_falling` by accounting for the presence of blocks
- all blocks are in contact with the platform floor at this point
- got rid of some entity flags that we don't need
- since we don't block enemies movement we got rid of the call

en_check_overlapping:
only sonic gets blocked (by setting xvel to zero), we simply reverse
the xvel for enemies if they are in contact with a block

en_check_falling:
in a nutshell an entity can fall from a block or from a platform,
the added code determines which one is the case by inspecting the
entity's block number `blockno`

en_apply_gravity:
here's where we implement wall-jumping and also where we set the
entity block number `blockno` to know later that the entity is on
a block

en_check_blocked:
just moves the entity a bit away from the block so that the
entity won't be blocked again (which is not what we want of course);
note that the player must have pressed the `tab-key` to make sonic
jump and so this has the effect of pushing sonic back (this happens
when `en_apply_gravity` gets called). This is another way to make
sonic do wall jumping on a block.

en_update_sonic:
we needed to apply a limit to sonic's position because we can now
move in the -x-direction and so if we are not careful sonic will
move beyond the limits of the level