impls game auto-mode
NOTES
related to task #22

https://github.com/misael-diaz/sonic-infinite-scroller-game-dev/issues/22

we can see the game from the perspective of the enemy in some sense

the game controls require little modification, all that we need to take
into account is that if we get arrow key input we interpret that as the
user wanting to control the camera instead of follow an entity. As in
the view-mode the e-key is used to cycle among enemy entities.

this feature did not require as much modification as the game view-mode
because the only thing that we are changing here is the entity that
we are looking at while the game runs; the switch from auto to normal
game mode does not require us to do anything special as for
transitioning from view mode to normal mode (restore the global frame
counter so that acrobatics do not break).